pyproject.toml
[tool.poetry]
name = "lms-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
ipython = "^8.23.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django = "4.2.1"
stripe = "^9.5.0"
redis = "^5.0.4"
celery = "^5.4.0"
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"
gevent = "^24.2.1"


[tool.poetry.group.lint.dependencies]
black = "^24.4.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.isort]
line_length = 119



[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.tox
    | \.venv
    | dist
  )/
)
'''
